requirements.txt
# Main packages
Flask~=2.1.1
flask-restx
Flask-CORS
Flask-SQLAlchemy
marshmallow~=3.15.0
PyJWT
SQLAlchemy~=1.4.35

# For test
pytest~=7.1.1
pytest-cov
